Move settings to webapp directory
